requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dateutil>=2.9.0
pytest>=8.0.0

